cmds/core/netcat: temporarily move listenForConnections() to new source
Move the listenForConnections() function, without any changes, to the new
source file "listen_conns.go", temporarily. Subsequently, I'll rewrite
listenForConnections() quite fundamentally, and for viewing that, mixing
the old hunks with the new would in fact be a detriment. By splitting
listenForConnections() out now, the rewrite will occur across files, and
both the old and the new versions of the function will be shown
self-contained.

Best view this patch with "git show --color --color-moved=zebra".
